third_party/blink/web_tests/external/wpt/mediacapture-streams/enumerateDevices-with-navigation.https-expected.txt

This is a testharness.js-based test.
[FAIL] enumerateDevices with navigation
  assert_equals: Promise state while inactive expected "pending" but got "rejected"
Harness: the test ran to completion.
